refactor: move bd division into ext method
